Create function for summarizing adwords event logs.
This creates a function can be used in a dplyr chain to summarize adwords event logs.
The function creates a table with a summary of relevant adwords info such as impressions,
clicks, CTR, Conversion Rate, ROAS.  Adding new columns based on an adwords elog is now
easy and the result can use "select()" to pull out a subset of columns.