Data frame creation to use summarize_adwords_elog function.
Use the newly created summarize_adwords_elog function to create the summaries.
This cleans up the code base and makes it easier to read.

This patch also moves the views to the end, so when the script runs it is easy
to skip Viewing the tables if it's not necessary.