Update search impression share calculations.
The search impression share and lost impression share
vectors were being calculated as though they were
represented the same, but they were represented differently.

That is, in the adwords data, lost share is capped at 90%,
but actual share is floored at 10%.

Now, they are calculated appropriately depending on if they are the
impression share or lost share metric.  The metric is also added to
other views where it is useful.