Finish implementing pulling Mixpanel data.
With this commit, there is now an alternative way to get data from mixpanel,
aside from pushing to the datawarehouse, waiting for it to populating, then joining.

Now the data is pulled from mixpanel and the datawarehouse separately, and
joined in R code.

Still need to integrate this functionality into the script, rather that kept in a test
file.