Refactor code to use mixpanel to get transaction data.
In previous versions, we were relying on an external script to pull from
Mixpanel and then push to the data warehouse, and then do a single [ugly] query
that pulls that data from the data warehouse all joined together.

This puts the power to pull data from the API directly in the hands of the R user.
As of this patch, the code is all in one script.  In future, it may make more sense
to pull out the data in a different "data retrieval" script that stores to CSV, then
read from a CSV, instead...mostly because a pull from Mixpanel takes a long time.